requirements.txt
pyfrc
robotpy-ctre
wpilib
pynetworktables
robotpy-wpilib-utilities